requirements.txt
librosa
numpy
soundfile
pydub
g2p-en
praat-parselmouth
torch
torchaudio
